# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MedCenter Back",
    "description": "API for MedCenter",
    "version": "1.0.0"
  },
  "host": "localhost:8088",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/create": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Создания анализа",
        "description": "Создание анализа для пациента",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Новость была создана",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "400": {
            "description": "Не удалось создать новость",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/list": {
      "post": {
        "tags": [
          "News"
        ],
        "summary": "Список новостей",
        "description": "Получения списка новостей по фильтрам",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Фильтры для получения новостей",
            "required": false,
            "schema": {
              "$ref": "#/definitions/NewsFilters"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/NewsList"
            },
            "description": "OK"
          },
          "400": {
            "schema": {
              "$ref": "#/definitions/NewsList"
            },
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": [
          "Analysis"
        ],
        "summary": "Получение списка анализов",
        "description": "Получение списка всех анализов с количеством",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Фильтры для получения анализов",
            "required": false,
            "schema": {
              "$ref": "#/definitions/AnalysFilters"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/AnalysisList"
            },
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Регистрация пользователя",
        "description": "Регистрация пользователя",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Пользовательские данные для регистрации",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AuthUserData"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Пользователь был зарегистрирован",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          },
          "400": {
            "description": "Не удалось зарегистрировать пользователя",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Авторизация пользователя",
        "description": "Авторизация пользователя с получением токенов",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Пользовательские данные для авторизации",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AuthUserData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь был успешно авторизован",
            "schema": {
              "type": "object",
              "properties": {
                "refs": {
                  "type": "string",
                  "example": "#/definitions/TokensData"
                }
              },
              "required": [
                "refs"
              ],
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Не удалось аутентифицировать пользователя",
            "schema": {
              "type": "object",
              "properties": {
                "refs": {
                  "type": "string",
                  "example": "#/definitions/Message"
                }
              },
              "required": [
                "refs"
              ],
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/change-password": {
      "patch": {
        "tags": ["Auth"],
        "summary": "Смена пароля",
        "description": "Смена пароля авторизованного пользователя. Требуется Bearer token.",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": { "new_password": { "type": "string", "example": "НовыйПароль123" } }
            }
          }
        ],
        "responses": {
          "200": { "description": "Пароль успешно изменён", "schema": { "$ref": "#/definitions/Message" } },
          "400": { "description": "Не удалось изменить пароль", "schema": { "$ref": "#/definitions/Message" } },
          "401": { "description": "Требуется авторизация", "schema": { "$ref": "#/definitions/Message" } }
        }
      }
    },
    "/me": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Проверка аутентифицированного пользователя",
        "description": "Проверка аутентифицированного пользователя путем декодирования токена",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Пользователь был аутентифицирован",
            "schema": {
              "type": "object",
              "properties": {
                "refs": {
                  "type": "string",
                  "example": "#/definitions/VerifyUser"
                }
              },
              "required": [
                "refs"
              ],
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Пользователь не прошел аутентификацию",
            "schema": {
              "type": "object",
              "properties": {
                "refs": {
                  "type": "string",
                  "example": "#/definitions/VerifyUser"
                }
              },
              "required": [
                "refs"
              ],
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Message": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Сообщение"
        }
      }
    },
    "VerifyUser": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "test_temail@mail.ru"
        }
      }
    },
    "AuthUserData": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "test@mail.ru"
        },
        "password": {
          "type": "string",
          "example": "test_user_password"
        }
      }
    },
...
    "NewsList": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "example": "Заголовок новости"
              },
              "description": {
                "type": "string",
                "example": "Описание новости"
              },
              "type": {
                "default": "GENERAL",
                "type": "string",
                "enum": [
                  "GENERAL",
                  "MED",
                  "SOCIAL",
                  "EVENTS",
                  "CONFERENCE",
                  "PERSONAL"
                ]
              },
              "createdAt": {
                "type": "string",
                "example": "24-12-2025"
              },
              "updatedAt": {
                "type": "string",
                "example": "24-12-2025"
              }
            }
          }
        },
        "total": {
          "type": "number",
          "example": 1
        }
      }
    },
    "NewsFilters": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Заголовок поиска"
        },
        "description": {
          "type": "string",
          "example": "Описание поиска"
        },
        "type": {
          "default": "GENERAL",
          "type": "string",
          "enum": [
            "GENERAL",
            "MED",
            "SOCIAL",
            "EVENTS",
            "CONFERENCE",
            "PERSONAL"
          ]
        },
        "createdAt": {
          "type": "string",
          "example": "24-12-2025"
        },
        "updatedAt": {
          "type": "string",
          "example": "24-12-2025"
        }
      }
    },
...
    "AnalysisList": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "Тип"
              },
              "text": {
                "type": "string",
                "example": "Текст"
              },
              "status": {
                "type": "string",
                "example": "DEAD"
              },
              "assignedDate": {
                "type": "string",
                "example": "2025-10-10"
              },
              "takenDate": {
                "type": "string",
                "example": "2025-10-10"
              },
              "readyDate": {
                "type": "string",
                "example": "2025-10-10"
              },
              "results": {
                "type": "object",
                "properties": {
                  "blood": {
                    "type": "string",
                    "example": "C+"
                  }
                }
              },
              "costs": {
                "type": "number",
                "example": 200
              },
              "updatedAt": {
                "type": "string",
                "example": "2025-10-11"
              },
              "createdAt": {
                "type": "string",
                "example": "2025-10-11"
              }
            }
          }
        },
        "total": {
          "type": "number",
          "example": 1
        }
      }
    },
    "AnalysFilters": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "кровь"
        },
        "costs": {
          "type": "number",
          "example": 1025
        },
        "status": {
          "type": "string",
          "example": "HEALTHY"
        }
      }
    },
...
  }
}